[client] hmm, instead of ProteanHax just go back to what we had before
It still leaks though. Or at least, it seems to. If I make them properties and
release in dealloc, UIStatusBarItemView tries to release them or something and
that crashes. But Instruments says it’s leaking? No clue what’s going on here
honestly, and every time I’ve tried to do what I thought was right with this
situation it always breaks in some way (autoreleasing eventually causes a bad
access and crashes; just setting them to nil breaks Protean).